+++ serverless/serverless.yml
@@ -41,6 +41,9 @@ functions:
     handler: api/hellobutton.echo
     memorySize: 128
     description: Hello Button develop echo
+    environment:
+      DB_HOST: hf-dev.c1dsfak07rhb.ap-northeast-2.rds.amazonaws.com
+      DB_NAME: hellobell_v2
     events:
       - http:
           path: api/echo
